requirements.txt
streamlit>=1.35.0
openai>=1.30.0
groq>=0.9.0
google-generativeai>=0.7.0
sentence-transformers>=2.7.0
numpy>=1.26.0
requests>=2.31.0
pdfplumber>=0.11.0
python-docx>=1.1.0
